requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
python-doctr>=0.6.0
matplotlib>=3.4.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
fuzzywuzzy>=0.18.0
gunicorn>=20.1.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
python-Levenshtein==0.12.2
pydantic==1.8.2
